requirements.txt
# hid requires external dependency hidapi
# see https://github.com/apmorton/pyhidapi
hid

pyyaml
pafy
python-vlc
youtube-dl

six